requirements_rk3588.txt
# RK3588 板端
rknn-toolkit-lite2
opencv-python
numpy
Pillow
PySide6
